refactor: move definitions from Haspara.Internal.* to Haspara.*
